refactor: 프로젝트 루트 이름 변경
