+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'seven'
+rootProject.name = 'cineAI'
