+++ src/main/java/com/amcamp/cineAI/CineAIApplication.java
@@ -1,12 +1,12 @@
-package com.amcamp.seven;
+package com.amcamp.cineAI;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SevenApplication {
+public class CineAIApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(SevenApplication.class, args);
+        SpringApplication.run(CineAIApplication.class, args);
     }
 }
