+++ src/main/java/com/amcamp/cineAI/global/config/swagger/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.amcamp.seven.global.config.swagger;
+package com.amcamp.cineAI.global.config.swagger;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
@@ -13,7 +13,7 @@ public class SwaggerConfig {
         Info info =
                 new Info()
                         .version("v0.0.1")
-                        .title("Seven API") // 이름
+                        .title("CineAI API") // 이름
                         .description("사용자 취향 맞춤형 영화 추천 프로젝트 API"); // 설명
         return new OpenAPI().info(info);
     }
